fixed -Wextra and -Werror flags

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: shurtado <shurtado@student.42barcelona.fr  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/03 04:55:12 by shurtado          #+#    #+#             */
-/*   Updated: 2024/08/01 15:35:59 by shurtado         ###   ########.fr       */
+/*   Updated: 2024/08/01 17:05:45 by shurtado         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ char	*get_next_line(int fd);
 int		ft_strlen(const char *s);
 char	*ft_strjoin(char const *s1, char const *s2);
 char	*ft_strdup(const char *s);
-char	*ft_substr(char const *s, unsigned int start, size_t len);
+char	*ft_substr(char const *s, unsigned int start, int len);
 char	*ft_strchr(const char *s, int c);
 
 #endif
